Benchmarks ported to OpenCRun
The 'opencrun' folder has been copied from the 'opencl' folder and it
contains the same OpenCL benchmarks with their source code and 'make'
files tailored for the OpenCRun runtime. The source code of the OpenCRun
benchmarks, for example, explicitly selects the CPU device type as its
the only one supported.